# Fireblocks.swagger.yml
...
openapi: 3.0.0
info:
  title: Fireblocks API
  version: "1.5.0"
servers:
  - url: 'https://api.fireblocks.io/v1'
paths:
...
  '/vault/accounts/{vaultAccountId}/{assetId}':
    get:
      summary: Returns a wallet of a specific asset inside a Fireblocks Vault account
      parameters:
        - in: path
          name: vaultAccountId
          required: true
          description: >-
            The ID of the vault account to return
          schema:
            type: string
        - in: path
          name: assetId
          required: true
          description: The ID of the asset
          schema:
            type: string
      responses:
        '200':
          description: A VaultAsset object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VaultAsset'
        default:
          $ref: '#/components/responses/Error'
    post:
      summary: Creates a new wallet in a Vault account
      parameters:
        - in: path
          name: vaultAccountId
          required: true
          description: >-
            The ID of the vault account to return, or 'default' for the default vault account
          schema:
            type: string
        - in: path
          name: assetId
          required: true
          description: The ID of the asset
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                eosAccountName:
                  description: Optional - when creating an EOS wallet, the account name. If not provided, a random name will be generated
                  type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreateVaultAssetResponse'
        default:
          $ref: '#/components/responses/Error'
  '/vault/accounts/{vaultAccountId}/{assetId}/balance':
    post:
      summary: Update balance and returns a wallet of a specific asset inside a Fireblocks Vault account
      parameters:
        - in: path
          name: vaultAccountId
          required: true
          description: >-
            The ID of the vault account to return
          schema:
            type: string
        - in: path
          name: assetId
          required: true
          description: The ID of the asset
          schema:
            type: string
      requestBody:
        required: false
        content:
          '*/*':
            schema:
              type: object
      responses:
        '200':
          description: A VaultAsset object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VaultAsset'
        default:
          $ref: '#/components/responses/Error'
...
security:
  - bearerTokenAuth: [ ]
  - ApiKeyAuth: [ ]
components:
  responses:
    Error:
      description: Error Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    CreateVaultAssetResponse:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        legacyAddress:
          type: string
        enterpriseAddress:
          type: string
        tag:
          type: string
        eosAccountName:
          type: string
        status:
          type: string
        activationTxId:
          type: string
    AllocatedBalance:
      type: object
      properties:
        allocationId:
          type: string
        thirdPartyAccountId:
          type: string
        affiliation:
          type: string
        virtualType:
          type: string
        total:
          type: string
        available:
          type: string
        pending:
          type: string
        frozen:
          type: string
        locked:
          type: string
        staked:
          type: string
    VaultAsset:
      type: object
      properties:
        id:
          type: string
        total:
          description: The total wallet balance. In EOS this value includes the network balance, self staking and pending refund. For all other coins it is the balance as it appears on the blockchain.
          type: string
        balance:
          deprecated: true
          description: Deprecated - replaced by "total"
          type: string
        available:
          description: Funds available for transfer. Equals the blockchain balance minus any locked amounts
          type: string
        pending:
          description: The cumulative balance of all transactions pending to be cleared
          type: string
        frozen:
          description: The cumulative frozen balance
          type: string
        lockedAmount:
          description: Funds in outgoing transactions that are not yet published to the network
          type: string
        staked:
          description: Staked balance
          type: string
        totalStakedCPU:
          type: string
        totalStakedNetwork:
          type: string
        selfStakedCPU:
          type: string
        selfStakedNetwork:
          type: string
        pendingRefundCPU:
          type: string
        pendingRefundNetwork:
          type: string
        blockHeight:
          type: string
        blockHash:
          type: string
          nullable: true
        allocatedBalances:
          type: array
          items:
            $ref: '#/components/schemas/AllocatedBalance'
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: number
          format: decimal
...